Fixing naming and move the structs to the moder.rs file
